requirements.txt
Django==2.0.7
django-leaflet==0.24.0
djangorestframework==3.8.2
leaflet==0.0.3
mysqlclient==1.3.13
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
